Update to 1.9.6
