+++ sample/proguard-rules.pro
@@ -21,5 +21,13 @@
 
 -keep public class * implements com.bumptech.glide.module.GlideModule;
 
+-keep class com.google.**
+-keep class autovalue.shaded.com.google.**
+-keep public class com.android.vending.licensing.ILicensingService
+
+-dontwarn org.apache.**
+-dontwarn com.google.**
+-dontwarn autovalue.shaded.com.google.**
+
 -dontwarn
 -ignorewarnings
