+++ sample/src/main/res/drawable/dashboard/drawable/ic_about.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M17.7,15.3H6.3v-0.9c0-1.9,3.8-3,5.7-3s5.7,1,5.7,3
+M12,5.1c1.4,0,2.6,1.1,2.6,2.6s-1.1,2.6-2.6,2.6S9.4,9.1,9.4,7.7 S10.6,5.1,12,5.1
+M18.7,2H5.3C4.3,2,3.4,2.8,3.4,3.9v13.3c0,1.1,0.9,1.9,1.9,1.9h3.8L12,22l2.9-2.9h3.8c1.1,0,1.9-0.9,1.9-1.9V3.9
+C20.6,2.8,19.7,2,18.7,2z" />
+</vector>
