+++ sample/src/main/res/drawable/dashboard/drawable/ic_android.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M14.8,8.9c-0.5,0-0.9-0.4-0.9-0.9S14.3,7,14.8,7s0.9,0.4,0.9,0.9S15.3,8.9,14.8,8.9
+M9.2,8.9c-0.5,0-0.9-0.4-0.9-0.9 S8.7,7,9.2,7s0.9,0.4,0.9,0.9S9.7,8.9,9.2,8.9
+M15.9,4.5l1.6-1.6c0.2-0.2,0.2-0.6,0-0.8v0c-0.2-0.2-0.6-0.2-0.8,0l-1.8,1.8
+C14,3.5,13,3.2,12,3.2c-1.1,0-2,0.3-2.9,0.7L7.3,2.2c-0.2-0.2-0.6-0.2-0.8,0l0,0c-0.2,0.2-0.2,0.6,0,0.8l1.6,1.6
+C6.5,5.7,5.4,7.6,5.4,9.8v0.9h13.1V9.8C18.6,7.6,17.5,5.7,15.9,4.5
+M5.4,15.4c0,3.6,2.9,6.6,6.6,6.6c3.6,0,6.6-2.9,6.6-6.6v-3.8H5.4 V15.4z" />
+</vector>
