+++ sample/src/main/res/drawable/dashboard/drawable/ic_apply.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#ffffff"
+        android:pathData="M12,9.8l-4.4,4.4h3.3v6.7h2.2v-6.7h3.3
+M19.8,3.1H4.2C3,3.1,2,4.1,2,5.3v13.3c0,1.2,1,2.2,2.2,2.2h4.4v-2.2H4.2V7.6h15.6
+v11.1h-4.4v2.2h4.4c1.2,0,2.2-1,2.2-2.2V5.3C22,4.1,21,3.1,19.8,3.1z" />
+</vector>
