+++ sample/src/main/res/drawable/dashboard/drawable/ic_code.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M11,2.5H13v1.9h0.5c0.8,0,1.4,0.6,1.4,1.4v3.4l-0.4,0.4l1.6,2.7c1.1-1.1,1.7-2.5,1.7-4.1h1.9c0,2.3-1,4.4-2.7,5.8l3,5.2
+l0.1,2.1l-1.8-1.2l-3-5.1c-1,0.5-2.2,0.8-3.4,0.8s-2.4-0.3-3.4-0.8l-3,5.1l-1.8,1.2l0.1-2.1l5.6-9.7L9.1,9.2V5.9
+c0-0.8,0.6-1.4,1.4-1.4H11V2.5
+M9.5,13.5c0.8,0.4,1.6,0.5,2.5,0.5s1.7-0.2,2.5-0.5l-1.4-2.4h0c-0.6,0.6-1.5,0.6-2.1,0h0L9.5,13.5
+M12,6.4c-0.5,0-1,0.4-1,1s0.4,1,1,1s1-0.4,1-1S12.5,6.4,12,6.4z" />
+</vector>
