+++ sample/src/main/res/drawable/dashboard/drawable/ic_copyright.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+        android:width="24dp"
+        android:height="24dp"
+        android:viewportWidth="24"
+        android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M10.08 10.86c.05-.33 .16 -.62 .3 -.87s.34-.46 .59 -.62c.24-.15 .54 -.22 .91 -.23
+.23 .01 .44 .05 .63 .13 .2 .09 .38 .21 .52 .36s.25 .33 .34 .53 .13 .42 .14 .64
+h1.79c-.02-.47-.11-.9-.28-1.29s-.4-.73-.7-1.01-.66-.5-1.08-.66-.88-.23-1.39-.23c-.65
+0-1.22 .11 -1.7 .34 s-.88 .53 -1.2 .92 -.56 .84 -.71 1.36S8 11.29 8 11.87v.27c0
+.58 .08 1.12 .23 1.64s.39 .97 .71 1.35 .72 .69 1.2 .91 1.05 .34 1.7 .34 c.47 0
+.91-.08 1.32-.23s.77-.36 1.08-.63 .56 -.58 .74 -.94 .29 -.74 .3 -1.15h-1.79c-.01
+.21 -.06 .4 -.15 .58 s-.21 .33 -.36 .46 -.32 .23 -.52 .3 c-.19 .07 -.39 .09 -.6
+.1
+-.36-.01-.66-.08-.89-.23-.25-.16-.45-.37-.59-.62s-.25-.55-.3-.88-.08-.67-.08-1v-.27c0-.35
+.03 -.68 .08 -1.01zM12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2
+12 2zm0 18c-4.41 0-8-3.59-8-8s3.59-8 8-8 8 3.59 8 8-3.59 8-8 8z" />
+</vector>
