+++ sample/src/main/res/drawable/dashboard/drawable/ic_libraries.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M12,7.8c1.6,0,2.9-1.3,2.9-2.9S13.6,2,12,2S9.1,3.3,9.1,4.9S10.4,7.8,12,7.8
+M12,11.3C9.7,9.2,6.6,7.8,3.2,7.8v10.7
+c3.4,0,6.5,1.3,8.8,3.4c2.3-2.1,5.4-3.4,8.8-3.4V7.8C17.4,7.8,14.3,9.2,12,11.3z" />
+</vector>
