+++ sample/src/main/res/drawable/dashboard/drawable/ic_no_connection.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M3.8,4.9l2.5,2.5C3.3,7.6,1,10,1,12.9c0,3,2.5,5.5,5.5,5.5h10.8l1.8,1.8l1.2-1.2L4.9,3.8
+M18.7,9.3
+c-0.6-3.2-3.4-5.5-6.7-5.5c-1.4,0-2.6,0.4-3.7,1.1l12.7,12.7c1.2-0.8,1.9-2.2,1.9-3.7C23,11.4,21.1,9.5,18.7,9.3z" />
+</vector>
