+++ sample/src/main/res/drawable/dashboard/drawable/ic_search.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#ffffff"
+        android:pathData="M9.8,3.5c3.5,0,6.3,2.8,6.3,6.3c0,1.6-0.6,3-1.5,4.1l0.3,0.3h0.8l4.9,4.9L19,20.5l-4.9-4.9v-0.8l-0.3-0.3
+c-1.1,0.9-2.5,1.5-4.1,1.5c-3.5,0-6.3-2.8-6.3-6.3S6.3,3.5,9.8,3.5
+M9.8,5.4c-2.4,0-4.4,1.9-4.4,4.4s1.9,4.4,4.4,4.4
+s4.4-1.9,4.4-4.4S12.2,5.4,9.8,5.4z" />
+</vector>
