+++ sample/src/main/res/drawable/dashboard/drawable/ic_translate.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+
+    <path
+        android:fillColor="#000000"
+        android:pathData="M12.8,14.8l-2.3-2.3l0,0c1.6-1.8,2.7-3.8,3.4-5.9h2.7V4.7h-6.4V2.9H8.4v1.8H2v1.8h10.2c-0.6,1.7-1.6,3.4-2.9,4.9
+c-0.8-0.9-1.5-2-2.1-3H5.4c0.7,1.5,1.6,2.9,2.7,4.1l-4.6,4.6l1.3,1.3l4.5-4.5l2.8,2.8L12.8,14.8
+M17.9,10.2h-1.8L12,21.1h1.8l1-2.7 h4.3l1,2.7H22L17.9,10.2
+M15.5,16.5l1.5-3.9l1.5,3.9H15.5z" />
+</vector>
