+++ sample/src/main/res/values/jahir.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright (c) 2016.  Jahir Fiquitiva
+  ~
+  ~ Licensed under the CreativeCommons Attribution-ShareAlike
+  ~ 4.0 International License. You may not use this file except in compliance
+  ~ with the License. You may obtain a copy of the License at
+  ~
+  ~    http://creativecommons.org/licenses/by-sa/4.0/legalcode
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ Big thanks to the project contributors. Check them in the repository.
+  ~
+  -->
+
+<resources xmlns:tools="http://schemas.android.com/tools" tools:ignore="MissingTranslation">
+
+	<bool name="dev_options">true</bool>
+	
+	<string-array name="nonconsumable_google_donation_items">
+		<item>soda</item>
+		<item>coffee</item>
+		<item>burger</item>
+		<item>pizza</item>
+		<item>meal</item>
+		<item>bigmeal</item>
+		<item>ram</item>
+		<item>device</item>
+		<item>quarterofuniversitysemester</item>
+    </string-array>
+
+    <string-array name="consumable_google_donation_items">
+		<item>soda</item>
+		<item>coffee</item>
+		<item>burger</item>
+		<item>pizza</item>
+		<item>meal</item>
+		<item>bigmeal</item>
+		<item>ram</item>
+		<item>device</item>
+		<item>quarterofuniversitysemester</item>
+    </string-array>
+
+    <string-array name="google_donations_catalog">
+		<item>Soda</item>
+		<item>Coffee</item>
+		<item>Burger</item>
+		<item>Pizza</item>
+		<item>Meal</item>
+		<item>Big Meal</item>
+		<item>New RAM</item>
+		<item>New Device</item>
+		<item>A Quarter Of University Semester</item>
+    </string-array>
+
+</resources>
