+++ sample/src/main/res/xml/themeinfo.xml
@@ -23,7 +23,7 @@
 
 <!-- This is for Atom Launcher -->
 <themeInfo>
-    <packageName>jahirfiquitiva.apps.iconshowcase.sample</packageName>
+    <packageName>jahirfiquitiva.apps.iconshowcase.sampleip</packageName>
     <versionName>1.0</versionName>
     <versionCode>1</versionCode>
     <themeName>IconShowcase Dashboard</themeName>
